move binaries for iBasso DX50 & DX90 to internal drive (/sdcard)
This patchset changes three things:
* binaries, libs, all rockbox files AND user settings are now in the
  same folder on the fat32 partition in the folder "/mnt/sdcard/.rockbox"
* to have rockbox_dir in /.rockbox, the dualboot selector app now
  creates a symbolic link at every start to /mnt/sdcard/.rockbox
  Additionally, the boot selector app shows an error picture when no
  rockbox binary is in the internal drive.
* buildzip.pl renames the .codecs to libx.so

This changings will simplify the update procedure a lot. No more patching
of the OF will be required, only for the first installation.
This new folder structure is now very close to the folder structure of the
no-application target.

+++ tools/buildzip.pl
@@ -428,6 +428,18 @@ sub buildzip {
         find(find_copyfile(qr/.*\.codec/, abs_path("$temp_dir/codecs/")), 'lib/rbcodec/codecs');
     }
 
+    # iBasso DX Players are hosted on android therefore need the shared lib filename
+    if ($modelname =~ /ibassodx/){
+        opendir(DIR,"$temp_dir/codecs/") or die "can't open CODECS temp dir";
+        my @codecs = readdir(DIR);
+        closedir(DIR);
+        foreach my $codec (@codecs) {
+            if ($codec =~ /(.*)\.codec/) {
+                move("$temp_dir/codecs/$1.codec", "$temp_dir/codecs/lib$1.so");
+            }
+        }
+    }
+
     # remove directory again if no codec was copied
     rmdir("$temp_dir/codecs");
